update：admin模块移除模板缓存类，修改添加模板时的返回信息，移除varibles变量，添加mybatis异常处理类

+++ apps/push-admin-api-app/src/main/java/com/mhd/push/publicapi/service/MessageTemplateService.java
@@ -34,9 +34,9 @@ public interface MessageTemplateService extends IService<MessageTemplate> {
      * 新增模板信息
      *
      * @param messageTemplateSaveDTO 模板信息
-     * @return 结果
+     * @return 新增后的模板ID，失败时返回 null
      */
-    int insertTemplate(MessageTemplateSaveDTO messageTemplateSaveDTO);
+    Long insertTemplate(MessageTemplateSaveDTO messageTemplateSaveDTO);
 
     /**
      * 修改模板信息
